# Conflicts:
#	.gitmodules
#	base/allocator/partition_allocator/src/partition_alloc/partition_alloc_base/logging.cc
#	base/allocator/partition_allocator/src/partition_alloc/partition_alloc_config.h
#	base/enterprise_util_mac.mm
#	base/logging.cc
#	base/mac/launch_application.mm
#	base/mac/launch_services_spi.h
#	base/synchronization/waitable_event_apple.cc
#	base/threading/platform_thread_apple.mm
#	chrome/browser/resources/settings/about_page/about_page.html
#	chrome/browser/ui/cocoa/share_menu_controller.mm
#	chrome/browser/ui/views/frame/browser_non_client_frame_view_mac.mm
#	chrome/services/mac_notifications/mac_notification_provider_impl.mm
#	components/bookmarks/browser/bookmark_node_data_mac.mm
#	components/bookmarks/browser/bookmark_pasteboard_helper_mac.mm
#	components/live_caption/caption_util.cc
#	components/remote_cocoa/app_shim/menu_controller_cocoa_delegate_impl.mm
#	components/remote_cocoa/app_shim/native_widget_ns_window_bridge.mm
#	components/remote_cocoa/app_shim/select_file_dialog_bridge.mm
#	components/soda/BUILD.gn
#	components/update_client/crx_downloader_factory.cc
#	content/app_shim_remote_cocoa/web_contents_view_cocoa.mm
#	content/app_shim_remote_cocoa/web_drag_source_mac.mm
#	content/browser/accessibility/browser_accessibility_state_impl_mac.mm
#	content/browser/renderer_host/clipboard_host_impl.cc
#	device/fido/BUILD.gn
#	device/fido/mac/util.mm
#	gpu/ipc/service/image_transport_surface_overlay_mac.h
#	gpu/ipc/service/image_transport_surface_overlay_mac.mm
#	media/capture/video/apple/sample_buffer_transformer.cc
#	net/base/platform_mime_util_apple.mm
#	net/ssl/ssl_platform_key_mac.cc
#	remoting/host/mac/constants_mac.h
#	remoting/host/mac/permission_utils.h
#	sandbox/policy/mac/common.sb
#	sandbox/policy/mac/on_device_model_execution.sb
#	services/device/public/cpp/device_features.cc
#	ui/base/clipboard/clipboard.h
#	ui/base/clipboard/clipboard_android.cc
#	ui/base/clipboard/clipboard_android.h
#	ui/base/clipboard/clipboard_data.h
#	ui/base/clipboard/clipboard_mac.h
#	ui/base/clipboard/clipboard_mac.mm
#	ui/base/clipboard/clipboard_non_backed.cc
#	ui/base/clipboard/clipboard_non_backed.h
#	ui/base/clipboard/clipboard_ozone.cc
#	ui/base/clipboard/clipboard_ozone.h
#	ui/base/clipboard/clipboard_util_win.h
#	ui/base/clipboard/clipboard_win.cc
#	ui/base/clipboard/clipboard_win.h
#	ui/base/clipboard/scoped_clipboard_writer.cc
#	ui/base/clipboard/scoped_clipboard_writer.h
#	ui/base/clipboard/test/test_clipboard.cc
#	ui/base/clipboard/test/test_clipboard.h
#	ui/base/dragdrop/os_exchange_data_provider_mac.h
#	ui/base/dragdrop/os_exchange_data_provider_mac.mm
#	ui/display/mac/test/virtual_display_util_mac.mm

+++ content/browser/download/mhtml_generation_browsertest.cc
@@ -144,7 +144,7 @@ class MockWriterBase : public mojom::MhtmlFileWriter {
   void WriteDataToProducerPipe(
       mojo::ScopedDataPipeProducerHandle producer_pipe) {
     base::ScopedAllowBlockingForTesting allow_blocking;
-    uint32_t size = strlen(kTestData);
+    size_t size = strlen(kTestData);
     producer_pipe->WriteData(kTestData, &size, MOJO_WRITE_DATA_FLAG_NONE);
     producer_pipe.reset();
   }
@@ -345,9 +345,9 @@ class MHTMLGenerationTest : public ContentBrowserTest,
     ASSERT_TRUE(has_mhtml_callback_run())
         << "Unexpected error generating MHTML file";
 
-    // TODO(crbug.com/997408): Add tests which will let MHTMLGeneration manager
-    // fail during file write operation. This will allow us to actually test if
-    // we receive a bogus hash instead of a std::nullopt.
+    // TODO(crbug.com/40641976): Add tests which will let MHTMLGeneration
+    // manager fail during file write operation. This will allow us to actually
+    // test if we receive a bogus hash instead of a std::nullopt.
     EXPECT_EQ(std::nullopt, file_digest());
 
     // Skip well formedness check if explicitly disabled or there was a
@@ -565,7 +565,7 @@ IN_PROC_BROWSER_TEST_P(MHTMLGenerationTest, GenerateMHTMLInNonTempDir) {
 
 // Regression test for the crash/race from https://crbug.com/612098.
 //
-// TODO(crbug.com/959435): Flaky on Android.
+// TODO(crbug.com/41456635): Flaky on Android.
 #if BUILDFLAG(IS_ANDROID)
 #define MAYBE_GenerateMHTMLAndCloseConnection \
   DISABLED_GenerateMHTMLAndCloseConnection
@@ -594,7 +594,7 @@ IN_PROC_BROWSER_TEST_P(MHTMLGenerationTest,
   EXPECT_EQ(ReadFileSizeFromDisk(path), file_size());
 }
 
-// TODO(crbug.com/672313): Flaky on Windows.
+// TODO(crbug.com/41290169): Flaky on Windows.
 #if BUILDFLAG(IS_WIN)
 #define MAYBE_InvalidPath DISABLED_InvalidPath
 #else
@@ -681,7 +681,7 @@ IN_PROC_BROWSER_TEST_P(MHTMLGenerationTest, GenerateMHTMLIgnoreNoStore) {
   EXPECT_THAT(mhtml, ContainsRegex("Content-Location:.*/nostore.html"));
 }
 
-// TODO(crbug.com/615291): These fail on Android under some circumstances.
+// TODO(crbug.com/40470937): These fail on Android under some circumstances.
 #if BUILDFLAG(IS_ANDROID)
 #define MAYBE_ViewedMHTMLContainsNoStoreContent \
   DISABLED_ViewedMHTMLContainsNoStoreContent
